Added function to get area of circle
Rename operator_ex1.py to operator_1.py
